% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryalphadigits.R
\docType{data}
\name{binaryalphadigits}
\alias{binaryalphadigits}
\title{Binary Alphadigits}
\source{
\url{http://www.cs.nyu.edu/~roweis/data.html}
}
\description{
Binary 20x16 digits of "0" through "9" and capital "A" through "Z". 39 examples of each class.
}
\references{
From Simon Lucas' (sml@essex.ac.uk), Algoval system.
}
\keyword{data}
