% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diabetes.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Diabetes Data}
\format{

}
\source{
\url{https://doi.org/10.1007/978-1-4612-5098-2}
}
\description{
Data  on 145 non-obese patients collected at the Stanford Clinical Reseach Center to investigate "the relationship between
chemical subclinical and overt nonketotic diabetes.  The three primary variables used in the analysis and  presented [here] are glucose intolerance, insulin response
}
\details{
This is a dataset from the "Data" book by Andrews and Herzberg (1985) Chapter 36, pp. 215-220  Table 36.1
}
\references{
David F. Andrews and Agnes M. Herzberg (1985) 
  "Data: A Collection of Problems from Many Fields for the Student and Research Worker", 
  Springer, New York. DOI:10.1007/978-1-4612-5098-2
}
\author{
R.W. Oldford
}
\keyword{Andrews-Herzberg}
\keyword{medical}
