% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digits.R
\docType{data}
\name{digits}
\alias{digits}
\title{USPS Handwritten Digits}
\format{Data frame with one image per column.}
\source{
\url{http://www.cs.nyu.edu/~roweis/data.html}
}
\description{
8-bit 16x16 grayscale images of "0" through "9"; 1100 examples of each class.
}
\keyword{data}
