% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faces.R
\docType{data}
\name{faces}
\alias{faces}
\title{Olivetti Faces}
\format{Data frame with one image per column.}
\source{
\url{http://www.cs.nyu.edu/~roweis/data.html}
}
\description{
Grayscale faces 8 bit [0-255], a few images of several different people.
}
\details{
400 total images, 64x64 size.

From the Olivetti database at ATT.
}
\keyword{data}
