% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frey.R
\docType{data}
\name{frey}
\alias{frey}
\title{Frey Faces}
\format{Data frame with one image per column.}
\source{
\url{http://www.cs.nyu.edu/~roweis/data.html}
}
\description{
From Brendan Frey. Almost 2000 images of Brendan's face, taken from
sequential frames of a small video. Size: 20x28.
}
\keyword{data}
