% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordalphadigits.R
\docType{data}
\name{ordalphadigits}
\alias{ordalphadigits}
\title{Binary Alphadigits Isomap}
\format{
Object of class 'isomap'.
}
\description{
Binary Alphadigits Isomap
}
\details{
Dissimilarity object of class 'isomap'. Returned from:

\code{isomap(vegdist(binaryalphadigits), k=6)}.

Introduced simply to cache the results of this step so as to speed up demos.
}
\seealso{
\code{\link{binaryalphadigits}}
}
\keyword{images}
