% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordfrey.R
\docType{data}
\name{ordfrey}
\alias{ordfrey}
\title{Frey Faces Isomap}
\format{
Object of class 'isomap'.
}
\description{
Frey Faces Isomap
}
\details{
Dissimilarity object of class 'isomap' for Frey Faces data, created with:

\code{isomap(vegdist(t(frey), method="euclidean"),k = 12, ndim=6, fragmentedOK = TRUE)}

Introduced simply to cache the results of this step so as to speed up demos.
}
\seealso{
\code{\link{frey}}
}
\keyword{images}
