% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_data.R
\name{pkg_data}
\alias{pkg_data}
\title{Simple summary of data available in named packages}
\usage{
pkg_data(package = NULL)
}
\arguments{
\item{package}{A character vector giving the package(s) to look in for data sets, 
or NULL. By default, all packages in the search path are used, then the ‘data’ 
subdirectory (if present) of the current working directory.}
}
\description{
Simple summary of data available in named packages
}
\details{
Syntactic sugar wrapping call to \code{utils::data(package = package)} 
to return basic information on datasets in package.
No data are loaded by the call.
}
\examples{
pkg_data("loon.data")
}
\seealso{
data
}
