% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryalphadigits.R
\docType{data}
\name{binaryalphadigits}
\alias{binaryalphadigits}
\title{Binary Alphadigits}
\format{A data frame with 1,404 rows each representing an image (39 examples of 36 classes)
  and 320 variables taking binary values 1 or 0 (black =1 and white = 0 for each pixel value
  of a 20x16 alpha-numeric image).}
\source{
\url{http://www.cs.nyu.edu/~roweis/data.html}
}
\description{
Binary 20x16 digits of "0" through "9" and capital "A" through "Z". 39 examples of each class.
}
\references{
From Simon Lucas' (sml@essex.ac.uk), Algoval system.
}
\seealso{
\code{\link{ordalphadigits}}
}
\keyword{images}
