\name{binaryalphadigits}
\docType{data}
\alias{binaryalphadigits}
\title{Binary Alphadigits}
\description{
 Binary 20x16 digits of "0" through "9" and capital "A" through "Z". 39 examples of each class.

 From Simon Lucas' (sml@essex.ac.uk), Algoval system. 
}
\usage{binaryalphadigits}
\format{Data frame with one image per row.}
\source{\url{http://www.cs.nyu.edu/~roweis/data.html}}
\keyword{datasets}
