\name{digits}
\docType{data}
\alias{digits}
\title{USPS Handwritten Digits}
\description{
8-bit 16x16 grayscale images of "0" through "9"; 1100 examples of each class. 
}
\usage{digits}
\format{Data frame with one image per column.}
\source{\url{http://www.cs.nyu.edu/~roweis/data.html}}
\keyword{datasets}
