\name{faces}
\docType{data}
\alias{faces}
\title{Olivetti Faces}
\description{
Grayscale faces 8 bit [0-255], a few images of several different people.

400 total images, 64x64 size.

From the Oivetti database at ATT. 
}
\usage{faces}
\format{Data frame with one image per column.}
\source{\url{http://www.cs.nyu.edu/~roweis/data.html}}
\keyword{datasets}
