% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frey.R
\docType{data}
\name{frey}
\alias{frey}
\title{Frey Faces}
\format{
Data frame with 1,965 variables (one image per column) and 560 rows 
(the 560 greyscale values of a 20x28 image).
}
\source{
\url{http://www.cs.nyu.edu/~roweis/data.html}
}
\description{
1,965 images of Brendan Frey's face, taken from
sequential frames of a small video. Image size: 20x28.
}
\seealso{
\code{\link{ordfrey}}, \code{\link{faces}}
}
\keyword{images}
