\name{frey}
\docType{data}
\alias{frey}
\title{Frey Face}
\description{
From Brendan Frey. Almost 2000 images of Brendan's face, taken from sequential frames of a small video. Size: 20x28. 

}
\usage{frey}
\format{Data frame with one image per column.}
\source{\url{http://www.cs.nyu.edu/~roweis/data.html}}
\keyword{datasets}
