\name{ordalphadigits}
\docType{data}
\alias{ordalphadigits}
\title{Dissimilarity object of class 'isomap' for Binary Alphadigits data}
\description{
Dissimilarity object of class 'isomap'. Returned from:

\code{isomap(vegdist(binaryalphadigits), k=6)}.
}
\usage{ordalphadigits}
\format{Object of class 'isomap'.}
\keyword{datasets}
