\name{ordfrey}
\docType{data}
\alias{ordfrey}
\title{Dissimilarity object of class 'isomap' for Frey Faces data}
\description{



Dissimilarity object of class 'isomap'. Returned from:

\code{isomap(vegdist(t(frey), method="euclidean"),k = 12, ndim=6, fragmentedOK = TRUE)}

}
\usage{ordfrey}
\format{Object of class 'isomap'.}
\keyword{datasets}
